backend-app/requirements.txt
fastapi
uvicorn
prisma
pydantic
python-dotenv
email-validator
pyjwt
python-multipart
pypdf
langchain
langchain-google-genai
langchain_community
langchain-chroma


#Excel
unstructured 
openpyxl
pandas

#PowerPoint
python-magic
python-pptx

#Word
docx2txt

google-generativeai
